requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
openpyxl==3.1.5
PyJWT==2.10.1
rich
lxml
prompt_toolkit
termgraph
